// Frontend/src/api/model/ascentSicherung.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type AscentSicherung = typeof AscentSicherung[keyof typeof AscentSicherung];


export const AscentSicherung = {
  vorstieg: 'vorstieg',
  toprope: 'toprope',
  solo: 'solo',
} as const;