// Frontend/src/api/model/ascentStyle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type AscentStyle = typeof AscentStyle[keyof typeof AscentStyle];


export const AscentStyle = {
  onsight: 'onsight',
  flash: 'flash',
  redpoint: 'redpoint',
  pinkpoint: 'pinkpoint',
  toprope: 'toprope',
  hangdog: 'hangdog',
  attempt: 'attempt',
} as const;